Patch from Diego 'Flameeyes' Pettenò.
ruby-prof was not working on an AMD64 box, because it 
does some nasty assumptions on int being enough to carry pointers.  His patch changes this behaviour to use long for everything to be safe, and INT2NUM rather than INT2FIX, as one never knows which size the numbers are when dealing with pointers.

The thread id also changed to unsigned long so that it does not show negative 
values as Thread ID.

git-svn-id: http://ruby-prof.rubyforge.org/svn@78 5fb4a865-4516-0410-ae3b-a609ddf12040